loading data, find peaks with median filter, and calculate with neighboring peaks
load data from text files into numpy arrays. use a median filter to
find peaks. look at points surrounding peaks and replace with average
or with one side if at edge. can handle peaks multiple points wide as
well as peaks at edge of file. next step is turn it into a function
that can go through each of the files in a folder, if it finds peaks
corrects them as well as saving a copy of the old files into a folder
that it creates